Add actual pre-agg SQL gen tests
